Switch to ulid (#485)
Makes it so we dont have to mental gymnastics to ensure that ordering on
createdAt is consistent. Makes cursors simpler

## Why ulid?

Because the generated ids are monotonically sortable (K-sortable). This
means we can sort them. The possible security concern of leaking the
creation order of items doesn't really matter for our cases.

Interesting reads:
-
https://github.com/paralleldrive/cuid2?tab=readme-ov-file#the-contenders
- https://supabase.com/blog/choosing-a-postgres-primary-key