Refactor fixture generation (#817)
**Background**
I was creating new fixtures due to some changes that needed to be made
to the database schema. I found it very annoying to generate the ulids
manually, seeing that the ulids should be absolutely ordered based on
insert time. Ended up rewriting such that the ids are generated by the
database.

**Changes**
- Use functions to generate fixtures instead of hard-coded objects so
that you can inject FK-IDs from previous fixtures
- some helper functions to make the code a bit neater