Replace ESLint & Prettier with Biome (#821)
[biomejs.dev](https://biomejs.dev/) is a modern replacement for both
prettier+eslint written in Rust. It's really really fast, like 10x the
speed of eslint+prettier and it has a really extensive linting ruleset.

As you can see from the diff, the formatting options are 100% compatible
with what we already have.

If we want to, we can also turn on organizeImports which alphabetically
sorts the imports. I have not done that yet, since that would dilute the
diff with like 100 files changed.

Instead of spending two minutes waiting for lintfix, it now takes 2.5
seconds for all the code


![image](https://github.com/dotkom/monoweb/assets/42585241/8299a39a-20f3-4346-a2af-dd3c975644d0)

## CI Lint times

The crazy :crab: :crab: language speed can also be felt in the CI
pipeline. Here is before and after


![image](https://github.com/dotkom/monoweb/assets/42585241/9cc3c22c-7f3c-4634-b881-92545864e850)

![image](https://github.com/dotkom/monoweb/assets/42585241/d8ba02df-188d-48c8-b82b-1630262e6b40)