add uri decoder to migration folder path (#823)
Migrations won't run when the migrations folder path has any URL encoded
characters. This is a problem if the repo is inside a folder with a
space character (definitely not me 👀)

NOTE: The migration url path could still be an issue in the build when it can't find migration folder.