Add sample data importer and prototype for frontpage (#855)
<img width="1258" alt="image"
src="https://github.com/dotkom/monoweb/assets/16722718/01bbc611-2711-4e09-bf57-c36177640d79">

This is just a prototype, but I think having a prototype to work on and
improve is better than having just a bunch of JSON data. The splash here
is mostly for companies, but my thought is that we remove it when a user
is signed in (we don't need to explain what Online is to people who have
Onlineweb-accounts), and replace it with maybe like a single large
promoted event, or something other we want to showcase (interest groups,
announcements, etc...). In the fadderweeks we can also show a large
promo for our splash page (which we need to make btw).

The sample data importer needs to be fixed when events-POC is merged.