Fix user syncing issue (#921)
The syncing service was using the ow user id from user in auth0 to
decide if user exists in local db or not. However, stuff might happen
such that a local user has been created but id is not updated in auth0.
I'm not sure how this can happen but it just happened to me locally.
Using auth0 id to check if user exists locally is safer because they
will always be the same.

Also fixes some logging statements that were not using string
interpolation.