Don't reset invoicification form when submitting (#992)
Bekk representative was complaining that when you click submit on the
form seemingly all the form values disappear, making it seem like an
empty form was submitted. This PR removes the form reset on submit.
Since we redirect on success anyways, it is not necessary to reset the
form when successfull, and if the submit fails it is also IMO better to
keep the values.