Merge deployment workflow with checks
Since we want the Docker build to be a part of the pipeline, but also
don't wish to repeat the entire build definition in each file, we merge
the files.

All jobs now depend on the "check" job that runs lint/type-check and will
require that job to pass to build containers (to preserve the fail-fast
behavior)

There is now a switch to push and call the apply workflow based on
whether the job was triggered from main push.

In the future I plan to put the production part in here too.