Copy full turbo prune tree in Docker builds
Fixes MONO-369

The JSON-prune tree from `turbo prune` causes the lockfile to point to
files that don't exist, rendering the entire lockfile broken and pnpm
having to build a fresh lock. This means the pnpm-lock.yaml we have in
root is not used at all, and this can lead to build differences.

When building with the full pruned tree this no longer happens.