Execute tasks on JobExecutor concurrently
NOTE: Concurrently does of course not mean in parallel - we are working in Node after all, and we're not using worker_threads.

Instead of maintaining a running state for the job executor, we should instead maintain running states on each individual job. Instead of blocking on all executions, we can now fire off jobs as soon as their seen, and thanks to database consistency, we will (hopefully) not see the same job twice.

The execution of the task has been changed to use its own Prisma transaction for isolation, while keeping the task state changes outside of said transaction.